interview-analysis-report/pyproject.toml
[project]
name = "interview-analysis-report"
version = "0.1.0"
description = "면접 분석 리포트 생성 서비스"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.89",
    "celery>=5.6.3",
    "langchain-openai>=1.1.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.0",
    "structlog>=25.5.0",
]
